% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop_to_column}
\alias{crop_to_column}
\title{Append Cropped Areas From Images Vertically}
\usage{
crop_to_column(images, geometry)
}
\arguments{
\item{images}{object of class "magick-image" representing one or more images}

\item{geometry}{geometry string}
}
\value{
object of class "magick-image"
}
\description{
Append Cropped Areas From Images Vertically
}
