% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop_to_row}
\alias{crop_to_row}
\title{Append Cropped Areas From Images Horizontally}
\usage{
crop_to_row(images, geometry)
}
\arguments{
\item{images}{object of class "magick-image" representing one or more images}

\item{geometry}{geometry string}
}
\value{
object of class "magick-image"
}
\description{
Append Cropped Areas From Images Horizontally
}
