% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{dim.matrix_image}
\alias{dim.matrix_image}
\title{Dimension of a "Matrix Image"}
\usage{
\method{dim}{matrix_image}(x)
}
\arguments{
\item{x}{object of class "matrix_image", as returned by 
\code{\link{image_matrix}}}
}
\value{
vector of two integers (number of rows, number of columns)
}
\description{
Dimension of a "Matrix Image"
}
