% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{image_matrix}
\alias{image_matrix}
\title{Arrange Images in a Matrix}
\usage{
image_matrix(images, nrow = NULL, ncol = NULL, byrow = FALSE, dimnames = NULL)
}
\arguments{
\item{images}{object of class "magick" representing one or more images}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{byrow}{logical. Shall images be arranged along rows or along columns?}

\item{dimnames}{list of vectors of row and column names, respectively}
}
\description{
Arrange Images in a Matrix
}
