% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{list_to_magick}
\alias{list_to_magick}
\title{Convert List of image-magick Objects to image-magick Object}
\usage{
list_to_magick(x)
}
\arguments{
\item{x}{list of objects of class "image-magick"}
}
\value{
object of class "image-magick"
}
\description{
Convert List of image-magick Objects to image-magick Object
}
