% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{[.matrix_image}
\alias{[.matrix_image}
\title{Extract Rows and Columns from "Matrix Image"}
\usage{
\method{[}{matrix_image}(x, ...)
}
\arguments{
\item{x}{object of class "matrix_image"}

\item{\dots}{arguments passed to \code{\link{extract}}}
}
\value{
object of class "image-magick"
}
\description{
Extract Rows and Columns from "Matrix Image"
}
