% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{try_image_read}
\alias{try_image_read}
\title{Try to Read Images with image_read}
\usage{
try_image_read(path)
}
\arguments{
\item{path}{passed to \code{\link[magick]{image_read}}}
}
\value{
object of class "magick-image", containing all images that could
  be read successfully from the paths given in \code{path}
}
\description{
Run \code{\link[magick]{image_read}} for each element of \code{path}. If 
an error occurs, continue with the next element.
}
