% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{extract}
\alias{extract}
\title{Extract Horizontal or Vertical Stripes from Image}
\usage{
extract(image, i = NULL, j = NULL, widths = NULL, heights = NULL)
}
\arguments{
\item{image}{object of class "matrix_image"}

\item{i}{row indices}

\item{j}{column indices}

\item{widths}{column widths in pixels}

\item{heights}{row heights in pixels}
}
\description{
Extract Horizontal or Vertical Stripes from Image
}
